Media: Move createTransientMedia to media utilities
